refactoring OBFService class
